# encoding/json: prevent duplicate slicebytetostring